More realistic core hour adjustment?
Previously, if a job used 1.1G per core on a system with 1G per core,
the statistics doubled the number of core hours. This was because, in
the worst case, this could mean cores remaining idle as a 1G per core
job wouldn't fit in th 0.9G per core.

A more realistic measurement could possibly be a 1.1G/1G = 10% uplift in
the core hours of such a job, hoping that worst and best cases average
out. That's what this commit does.